entities: added EntityInterface and changed workings
This will require some explaination. Entities will not use functions
defined in entity.h anymore, instead, they will be controled via small
vtables/interfaces that is EntityInterface. This will also provide a new
way to develop systems like mobs, which the the entity system can
provide a facade to communicate with that entity.

For example: ent_take_damage() is only a facade for
EntityInterface.take_damage(), so the fireball entity can check if the
entity implements that interface and operate acordingly.